+++ core/src/test/java/com/ericsson/bss/cassandra/ecchronos/core/repair/state/TestRepairStateImpl.java
@@ -22,9 +22,10 @@ import com.ericsson.bss.cassandra.ecchronos.core.utils.LongTokenRange;
 import com.ericsson.bss.cassandra.ecchronos.core.utils.TableReference;
 import com.google.common.collect.ImmutableSet;
 import org.assertj.core.util.Lists;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mockito.ArgumentCaptor;
+import org.mockito.Captor;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
@@ -34,7 +35,6 @@ import java.util.List;
 import java.util.concurrent.TimeUnit;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.eq;
 import static org.mockito.Matchers.isNull;
 import static org.mockito.Mockito.mock;
@@ -63,11 +63,8 @@ public class TestRepairStateImpl
     @Mock
     private ReplicaRepairGroup mockReplicaRepairGroup;
 
-    @Before
-    public void init()
-    {
-        when(mockReplicaRepairGroupFactory.generateReplicaRepairGroups(any())).thenReturn(Collections.singletonList(mockReplicaRepairGroup));
-    }
+    @Captor
+    private ArgumentCaptor<List<VnodeRepairState>> repairGroupCaptor;
 
     @Test
     public void testInitialEmptyState()
@@ -86,7 +83,7 @@ public class TestRepairStateImpl
                 .build();
 
         when(mockVnodeRepairStateFactory.calculateNewState(eq(tableReference), isNull(RepairStateSnapshot.class))).thenReturn(vnodeRepairStates);
-        when(mockReplicaRepairGroupFactory.generateReplicaRepairGroups(any())).thenReturn(Lists.emptyList());
+        when(mockReplicaRepairGroupFactory.generateReplicaRepairGroups(repairGroupCaptor.capture())).thenReturn(Lists.emptyList());
 
         RepairState repairState = new RepairStateImpl(tableReference, repairConfiguration,
                 mockVnodeRepairStateFactory, mockHostStates,
@@ -94,6 +91,7 @@ public class TestRepairStateImpl
 
         RepairStateSnapshot repairStateSnapshot = repairState.getSnapshot();
 
+        assertThat(repairGroupCaptor.getValue()).isEmpty();
         assertRepairStateSnapshot(repairStateSnapshot, expectedAtLeastRepairedAt, Lists.emptyList(), vnodeRepairStates);
 
         verify(mockTableRepairMetrics).repairState(eq(tableReference), eq(1), eq(0));
@@ -119,6 +117,7 @@ public class TestRepairStateImpl
                 .build();
 
         when(mockVnodeRepairStateFactory.calculateNewState(eq(tableReference), isNull(RepairStateSnapshot.class))).thenReturn(vnodeRepairStates);
+        when(mockReplicaRepairGroupFactory.generateReplicaRepairGroups(repairGroupCaptor.capture())).thenReturn(Collections.singletonList(mockReplicaRepairGroup));
 
         RepairState repairState = new RepairStateImpl(tableReference, repairConfiguration,
                 mockVnodeRepairStateFactory, mockHostStates,
@@ -126,6 +125,12 @@ public class TestRepairStateImpl
 
         RepairStateSnapshot repairStateSnapshot = repairState.getSnapshot();
 
+        List<VnodeRepairState> capturedVnodeStates = repairGroupCaptor.getValue();
+
+        assertThat(capturedVnodeStates).hasSize(1);
+        VnodeRepairState capturedVnodeRepairState = capturedVnodeStates.get(0);
+
+        assertVnodeRepairStateRepairedBefore(vnodeRepairState, capturedVnodeRepairState, System.currentTimeMillis() - repairIntervalInMs);
         assertRepairStateSnapshot(repairStateSnapshot, expectedAtLeastRepairedAt, Collections.singletonList(mockReplicaRepairGroup), vnodeRepairStates);
 
         verify(mockTableRepairMetrics).repairState(eq(tableReference), eq(1), eq(1));
@@ -146,7 +151,7 @@ public class TestRepairStateImpl
                 .build();
 
         when(mockVnodeRepairStateFactory.calculateNewState(eq(tableReference), isNull(RepairStateSnapshot.class))).thenReturn(vnodeRepairStates);
-        when(mockReplicaRepairGroupFactory.generateReplicaRepairGroups(any())).thenReturn(Lists.emptyList());
+        when(mockReplicaRepairGroupFactory.generateReplicaRepairGroups(repairGroupCaptor.capture())).thenReturn(Lists.emptyList());
 
         RepairState repairState = new RepairStateImpl(tableReference, repairConfiguration,
                 mockVnodeRepairStateFactory, mockHostStates,
@@ -154,10 +159,11 @@ public class TestRepairStateImpl
 
         RepairStateSnapshot repairStateSnapshot = repairState.getSnapshot();
 
+        assertThat(repairGroupCaptor.getValue()).isEmpty();
         assertRepairStateSnapshot(repairStateSnapshot, expectedRepairedAt, Lists.emptyList(), vnodeRepairStates);
 
         verify(mockTableRepairMetrics).repairState(eq(tableReference), eq(1), eq(0));
-        verify(mockTableRepairMetrics).lastRepairedAt(eq(tableReference), eq(repairStateSnapshot.lastRepairedAt()));
+        verify(mockTableRepairMetrics).lastRepairedAt(eq(tableReference), eq(expectedRepairedAt));
         reset(mockTableRepairMetrics);
 
         // Perform update
@@ -180,6 +186,15 @@ public class TestRepairStateImpl
         assertThat(repairStateSnapshot.getRepairGroups()).isEqualTo(replicaRepairGroups);
         assertThat(repairStateSnapshot.canRepair()).isEqualTo(canRepair);
         assertThat(repairStateSnapshot.getVnodeRepairStates()).isEqualTo(vnodeRepairStatesBase.combineWithRepairedAt(repairStateSnapshot.lastRepairedAt()));
+    }
+
+    private void assertVnodeRepairStateRepairedBefore(VnodeRepairState baseVnodeRepairState, VnodeRepairState actualVnodeRepairState, long repairedBefore)
+    {
+        assertThat(actualVnodeRepairState.getTokenRange()).isEqualTo(baseVnodeRepairState.getTokenRange());
+        assertThat(actualVnodeRepairState.getReplicas()).isEqualTo(baseVnodeRepairState.getReplicas());
+
+        assertThat(actualVnodeRepairState.lastRepairedAt()).isNotEqualTo(VnodeRepairState.UNREPAIRED);
+        assertThat(actualVnodeRepairState.lastRepairedAt()).isLessThanOrEqualTo(repairedBefore);
     }
 
     private RepairConfiguration repairConfiguration(long repairIntervalInMs)