+++ CHANGES.md
@@ -7,6 +7,7 @@
 * Reduce memory footprint - Issue #54
 * Locking failures log too much - Issue #58
 * Fix log file rotation - Issue #61
+* Correct initial replica repair group - Issue #60
 
 ## Version 1.0.0